requirements.txt
poke-env
python-dotenv
openai
anthropic
google-generativeai
requests
tqdm
tabulate